Removing old files. Adding fixed utils
