+++ pyproject.toml.jinja
@@ -19,7 +19,6 @@ dependencies = [
 scripts = [
     "hydra-colorlog>=1.2.0",
     "hydra-core>=1.3.2",
-    "python-dotenv>=1.0.1",
 ]
 dev = [
     "pre-commit>=4.1.0",
