+++ tests/test_broken_utils.py.jinja
@@ -6,7 +6,7 @@ import os
 import numpy as np
 import pytest
 
-from {{project_slug}}.utils import add_one, add_two, patchify_images
+from {{project_slug}}.broken_utils import add_one, add_two, patchify_images
 
 @pytest.fixture(scope='function')  # https://docs.pytest.org/en/6.2.x/fixture.html#fixture-scopes
 def my_fixture():
